Converted to a right handed coordinate system
To avoid confusion I converted the coordinate system to a right-handed
system. Most conventions use a right-handed system so I changed the
system to be +x axis pointed to the right of the screen, +y axis pointed
to the bottom of the screen, and +z axis pointed into the screen.